+++ src/main/java/co/edu/practice/tse/collections/PublicMemory.java
@@ -18,9 +18,9 @@ import java.util.List;
 @Data
 public class PublicMemory extends PrivateMemory {
     private List<Visualization> visualizationList;
-    private int numViews;
+    private Long numViews;
 
-    public PublicMemory(String id, String name, LocalDate memoryDate, LocalDate creationDate, EnumVisibility visibility, List<String> tagList, String creatorId, List<MemoryPhoto> memoryPhotoList, Location location, List<Visualization> visualizationList, int numViews) {
+    public PublicMemory(String id, String name, LocalDate memoryDate, LocalDate creationDate, EnumVisibility visibility, List<String> tagList, String creatorId, List<MemoryPhoto> memoryPhotoList, Location location, List<Visualization> visualizationList, Long numViews) {
         super(id, name, memoryDate, creationDate, visibility, tagList, creatorId, memoryPhotoList, location);
         this.visualizationList = visualizationList;
         this.numViews = numViews;
